+++ tensorrt_llm/_torch/auto_deploy/compile/backends/__init__.py
@@ -1,3 +1,1 @@
-import importlib.util
-
-from . import torch_compile, torch_opt, torch_simple
+from . import torch_compile, torch_cudagraph, torch_opt, torch_simple
